Change to use p256 instead of p521
It seems the original choice to use P-521 may have been a mistake, and
in any case there's no real cryptographic justification for using it
here over P-256 which will be faster and more widely supported.
Specifically, we've heard that envoy struggles with P-521 so switching
will help with that use case.

This also removes the approver check for the curve. That check was
incorrect (just checking the bit size of the underlying field doesn't
confirm that the curve is P-521) and was not in any case a worthwhile
thing to confirm since we shouldn't really care what key is used.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>