remove issuerRef for approver
The approver currently filters CertificateRequest resources based on the
issuerRef they have configured, which will be incredibly complicated to
handle in a scenario where we allow runtime configuration of issuers.

It also doesn't really provide any guarantees, guardrails or value, but
does introduce a race condition when the issuer is updated. Updating the
issuer cannot be synchronized everywhere at the same time, so there's
always a period after an issuer update in which the driver could be
using the new issuer and the approver could be looking for the old
issuer.

This commit removes the issuerRef for the approver while keeping
backwards compatibility with existing uses of the `--issuer-*` flags.

It replaces filtering of CertificateRequests by issuerRef with filtering
based on the presence of the standard annotation which contains the
SPIFFE ID. This helps to ensure that irrelevant CR resources are not
considered.

Signed-off-by: Ashley Davis <ashley.davis@venafi.com>