tweak camanager testing
Because we set the update retry period to 5s, polling every second is
wasteful as that's 5x faster than the retry period.

That logic also implies that a timeout of 60s is very short, since
that's only at most 12 attempts for the update to succeed. Anecdotal
evidence shows that this leads to a lot of flaked tests.

This commit increases that timeout to reduce flakes.

It also changes how the test works to clean up after it modifies global
state, and to compare strings rather than byte slices which makes error
output much easier to read if there's a problem.

Signed-off-by: Ashley Davis <ashley.davis@venafi.com>